// mobile-rn-archive/services/poseDetection/types.ts
export enum ExerciseType {
  SQUAT = 'squat',
  DEADLIFT = 'deadlift',
  PUSH_UP = 'pushup',
  BENCH_PRESS = 'bench_press',
  OVERHEAD_PRESS = 'overhead_press'
}